requirements.txt
python-dotenv>=1.0.0
pyodbc>=4.0.0

